+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Lido public class [#0020](https://github.com/lidofinance/lido-python-sdk/pull/20)
 
+### Changed
+- If strict is false, for validation we will try possible WC only for already used keys [#0024](https://github.com/lidofinance/lido-python-sdk/pull/24)
+
 ## [0.3.0](https://github.com/lidofinance/lido-python-sdk) - 2021-10-26
 ### Fixed
 - Move eth2deposit code to repository