+++ CHANGELOG.md
@@ -6,6 +6,11 @@ This changelog format is based on [Keep a Changelog](https://keepachangelog.com/
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/lidofinance/lido-python-sdk)
+### Added
+- Lido public class [#0020](https://github.com/lidofinance/lido-python-sdk/pull/20)
+
+### Changed
+- If strict is false, for validation we will try possible WC only for already used keys [#0024](https://github.com/lidofinance/lido-python-sdk/pull/24)
 
 ## [0.3.0](https://github.com/lidofinance/lido-python-sdk) - 2021-10-26
 ### Fixed