+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project are documented in this file.
 This changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/lidofinance/lido-python-sdk)
+## [1.0.0](https://github.com/lidofinance/lido-python-sdk) - 2021-10-31
 ### Added
 - Lido public class [#0020](https://github.com/lidofinance/lido-python-sdk/pull/20)
 - Added fast BLS verification [#0019](https://github.com/lidofinance/lido-python-sdk/pull/19)